### 2025-03-15 #7904 @editorialbot
[PRE REVIEW]: The Explabox: Model-Agnostic Machine Learning Transparency & Analysis
**Submitting author:** @marcelrobeer (<a href="http://orcid.org/0000-0002-6430-9774">Marcel Robeer</a>)
**Repository:** https://github.com/MarcelRobeer/explabox
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
