### 2025-03-16 0ae1225
Removed conclusion from paper
Signed-off-by: Marcel Robeer <marcel_robeer@hotmail.com>
