+++ .github/workflows/check.yml
@@ -14,7 +14,7 @@ jobs:
     - uses: actions/checkout@v4
     - name: Get changed files
       id: changed_files
-      uses: tj-actions/changed-files@v43
+      uses: tj-actions/changed-files@v46
       with:
         files: |
           *.md
@@ -75,7 +75,7 @@ jobs:
     - uses: actions/checkout@v4
     - name: Get changed files
       id: changed_files_py
-      uses: tj-actions/changed-files@v43
+      uses: tj-actions/changed-files@v46
       with:
         files: |
           explabox
