### 2025-03-16 94bbe57
Added DOIs
Signed-off-by: Marcel Robeer <marcel_robeer@hotmail.com>
