+++ paper/paper.bib
@@ -2,6 +2,7 @@
     author = {Ribeiro, Marco Tulio and Singh, Sameer and Guestrin, Carlos},
     booktitle = {AAAI Conference on Artificial Intelligence, Proceedings},
     title = {{Anchors: High-Precision Model-Agnostic Explanations}},
+    doi = {10.1609/aaai.v32i1.11491},
     year = {2018}
 }
 @inproceedings{Ribeiro2016a,
@@ -12,6 +13,7 @@
     issn = {9781450321389},
     pages = {1135--1144},
     title = {{``Why Should I Trust You?'': Explaining the Predictions of Any Classifier}},
+    doi = {10.18653/v1/n16-3020},
     year = {2016}
 }
 @inproceedings{Lundberg2017,
@@ -26,6 +28,7 @@
 @inproceedings{Ribeiro2020,
     author = {Marco Tulio Ribeiro and Tongshuang Wu and Carlos Guestrin and Sameer Singh},
     title = {{Beyond Accuracy: Behavioral Testing of NLP models with CheckList}},
+    doi = {10.18653/v1/2020.acl-main.442},
     booktitle = {Association for Computational Linguistics (ACL)},
     year = {2020}
 }
@@ -265,12 +268,13 @@
 }
 
 @article{Veale2021,
-    title={{Demystifying the Draft EU Artificial Intelligence Act—Analysing the good, the bad, and the unclear elements of the proposed approach}},
+    title={{Demystifying the Draft EU Artificial Intelligence Act — Analysing the good, the bad, and the unclear elements of the proposed approach}},
     author={Veale, Michael and Zuiderveen Borgesius, Frederik},
     journal={Computer Law Review International},
     volume={22},
     number={4},
     pages={97--112},
+    doi={10.9785/cri-2021-220402},
     year={2021},
 }
 @article{Edwards2022,
