+++ paper/paper.md
@@ -65,22 +65,13 @@ The `Explabox` turns these *ingestibles* into *digestibles*: pieces of informati
 
 **Explain** uses model-agnostic techniques [@Ribeiro2016b] to explain model behavior (*global*) and individual predictions (*local*). It summarizes model-labelled data, through prototypes (`K-Medoids`) or prototypes and criticisms (`MMDCritic` [@Kim2016]), and token distributions (`TokenFrequency`) and informativeness (`TokenInformation`) for the text modality. Local explanations are given by popular techniques for feature attribution scores (`LIME` [@Ribeiro2016a], `KernelSHAP` [@Lundberg2017]), relevant feature subsets (`Anchors` [@Ribeiro2018]), local rule-based models (`LORE` [@Guidotti2018]), and counterfactual/contrastive explanations (`FoilTrees` [@Waa2018]). These methods are constructed from generic components that split the relevant steps in global and local explanation generation. This allows customizability and for scientific advancements to quickly end up in operational processes. For example, they may combine the data sampling from `KernelSHAP` and summarize these with a surrogate rule-based model provided by the `imodels` package [@Singh2021]. However, to ease users into adoption we provide example configurations, such as `LIME` with default hyperparameters. **Explain** is provided by subpackage `text_explainability` [@text_explainability], which doubles as a standalone tool.
 
-**Expose** gathers sensitivity insights through local/global testing regimes. These insights can be used to, through relevant attributes, assess the *robustness* (e.g., the effect of typos on model performance), *security* (e.g., if inputs containing certain characters crash the model) and *fairness* (e.g., subgroup performance for protected attributes such as country of origin, gender, race or socioeconomic status) of the model. Relevant attributes can either be observed in the current data or generated from user-provided templates [@Ribeiro2020]. These attributes are then either summarized in performance metrics, compared to expected behavior [@Ribeiro2020], or assessed with fairness metrics for classification [@Mehrabi2021] and regression [@Agarwal2019]. Like **explain**, **expose** is also made from generic components, which allows users to customize data generation and tests. **Expose** is provided by the `text_sensitivity` subpackage [@text_sensitivity], which also doubles as a standalone tool.
+**Expose** gathers sensitivity insights through local/global testing regimes. These insights can be used to, through relevant attributes, assess the *robustness* (e.g., the effect of typos on model performance), *security* (e.g., if inputs containing certain characters crash the model) and *fairness* (e.g., subgroup performance for protected attributes such as country of origin, gender, race or socioeconomic status) of the model. Relevant attributes can either be observed in the current data or generated from user-provided templates [@Ribeiro2020] filled with multi-language data generation [@Faker]. These attributes are then either summarized in performance metrics, compared to expected behavior [@Ribeiro2020], or assessed with fairness metrics for classification [@Mehrabi2021] and regression [@Agarwal2019]. Like **explain**, **expose** is also made from generic components, which allows users to customize data generation and tests. **Expose** is provided by the `text_sensitivity` subpackage [@text_sensitivity], which also doubles as a standalone tool.
 
 ## Digestibles
 
-To serve diverse stakeholders' needs&mdash;such as auditors, applicants, end-users or clients [@Tomsett2018]&mdash;in consuming model and data insights, the digestibles are accessible through different channels like an interactive user interface (UI) for Jupyter Notebook or webpages (\autoref{fig:ui}), an API for integration with other tooling, and static reporting.
[…]
