### 2025-03-17 #7904 @MarcelRobeer
Shortened the paper length by removing the 'Open source implementation' section

```@editorialbot generate pdf```
