### 2025-04-14 #7904 @abhishektiwari
@MarcelRobeer Following citations from `.bib` file are unused. Please purge them

```
Abdul2018
Henin2020
Li2022
Mitchell2019
Sundararajan2017
Toreini2020
Veale2021
```

Once done. Can you please check if you can find DOI for remaining references for which DOI check was skipped? I can find DOI for some of them. E.g. _A Survey on Bias and Fairness in Machine Learning_ - [10.1145/3457607](https://doi.org/10.1145/3457607)
