### 2025-04-18 6f2f251
Removed superfluous references, added DOI
