+++ paper/paper.bib
@@ -20,7 +20,6 @@
     author = {Lundberg, Scott M. and Lee, Su-In},
     booktitle = {Advances in Neural Information Processing Systems 30 (NIPS 2017)},
     issn = {10495258},
-    mendeley-groups = {PhD},
     pages = {4765--4774},
     title = {{A Unified Approach to Interpreting Model Predictions}},
     year = {2017}
@@ -32,27 +31,11 @@
     booktitle = {Association for Computational Linguistics (ACL)},
     year = {2020}
 }
-@inproceedings{Sundararajan2017,
-    author = {Sundararajan, Mukund and Taly, Ankur and Yan, Qiqi},
-    booktitle = {34th International Conference on Machine Learning (ICML), Proceedings},
-    issn = {1938-7228},
-    title = {{Axiomatic Attribution for Deep Networks}},
-    year = {2017}
-}
 @inproceedings{Waa2018,
     author = {van der Waa, Jasper and Robeer, Marcel and van Diggelen, Jurriaan and Neerincx, Mark and Brinkhuis, Matthieu},
     booktitle = {2018 Workshop on Human Interpretability in Machine Learning (WHI 2018)},
     title = {{Contrastive Explanations with Local Foil Trees}},
     year = {2018}
-}
-@article{Mitchell2019,
-    title={{Model Cards for Model Reporting}},
-    ISBN={9781450361255},
-    DOI={10.1145/3287560.3287596},
-    journal={Proceedings of the Conference on Fairness, Accountability, and Transparency},
-    publisher={ACM},
-    author={Mitchell, Margaret and Wu, Simone and Zaldivar, Andrew and Barnes, Parker and Vasserman, Lucy and Hutchinson, Ben and Spitzer, Elena and Raji, Inioluwa Deborah and Gebru, Timnit},
-    year={2019},
 }
 @article{Guidotti2018,
     title={{Local Rule-Based Explanations of Black Box Decision Systems}},
@@ -75,6 +58,7 @@
     number={6},
     pages={1--35},
     year={2021},
+    doi={10.1145/3457607},
     publisher={ACM New York, NY, USA}
 }
 @misc{instancelib,
@@ -108,60 +92,18 @@
     eprint = {1806.07552},
     year = {2018}
 }
-@inproceedings{Toreini2020,
-    author = {Toreini, Ehsan and Aitken, Mhairi and Coopamootoo, Kovila and Elliott, Karen and Zelaya, Carlos Gonzalez and van Moorsel, Aad},
-    title = {{The Relationship between Trust in AI and Trustworthy Machine Learning Technologies}},
-    year = {2020},
-    isbn = {9781450369367},
-    publisher = {Association for Computing Machinery},
-    address = {New York, NY, USA},
-    doi = {10.1145/3351095.3372834},
-    booktitle = {Proceedings of the 2020 Conference on Fairness, Accountability, and Transparency},
-    pages = {272–283},
-    numpages = {12},
-    keywords = {trustworthiness, trust, machine learning, artificial intelligence},
-    location = {Barcelona, Spain},
-    series = {FAT* '20}
-}
-@inproceedings{Abdul2018,
-    title = {{Trends and Trajectories for Explainable, Accountable and Intelligible Systems: An HCI Research Agenda}},
-    author = {Ashraf Abdul and Jo Vermeulen and Danding Wang and Lim, {Brian Y.} and Mohan Kankanhalli},
-    year = {2018},
-    month = apr,
-    day = {20},
-    doi = {10.1145/3173574.3174156},
-    language = {English},
-    isbn = {978-1-4503-5620-6},
-    series = {CHI '18},
-    pages = {582:1--582:18},
[…]
