### 2025-04-18 #7904 @MarcelRobeer
@abhishektiwari , thank you very much for reviewing our paper.

I have removed the superfluous references from the `.bib` file, and have added the DOI for for the paper `Mehrabi2021`. See the latest commit ([6f2f251](https://github.com/MarcelRobeer/explabox/commit/6f2f25128f85a895536b73d210295931e2f00c69)) for the changes.

Would you also like met to add in ACM DOIs (e.g. https://dl.acm.org/doi/10.5555/1953048.2078195 for scikit-learn) as a url? These are not strictly DOIs, but function similarly
