### 2025-04-19 #7904 @abhishektiwari
@MarcelRobeer Lets ignore ACM DOI strong, those are not DOI truly.

I still see wordcount for `paper.md` is **1303**. Can you please try to see if you can bring it down to <= 1000 words?
