### 2025-04-25 27ba0e0
Shortened paper length
