+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'The Explabox: Model-Agnostic Machine Learning Transparency & Analysis'
+title: 'Explabox: Model-Agnostic Machine Learning Transparency & Analysis'
 tags:
     - Python
     - explainable AI (XAI)
@@ -38,43 +38,43 @@ bibliography: paper.bib
 
 # Summary
 
-We present the `Explabox`: an open-source toolkit for transparent and responsible machine learning (ML) model development and usage. `Explabox` aids in achieving explainable, fair and robust models by employing a four-step strategy: *explore*, *examine*, *explain* and *expose*. These steps offer model-agnostic analyses that transform complex `ingestibles` (models and data) into interpretable `digestibles`. The toolkit encompasses digestibles for descriptive statistics, performance metrics, model behavior explanations (local and global), and robustness, security, and fairness assessments. Implemented in Python, `Explabox` supports multiple interaction modes and builds on open-source packages. It empowers model developers and testers to operationalize explainability, fairness, auditability, and security. The initial release focuses on text data and models, with plans for expansion. `Explabox`'s code and documentation are available open-source at [https://explabox.readthedocs.io](https://explabox.readthedocs.io/en/stable).
+`Explabox` is an open-source toolkit for transparent and responsible machine learning (ML) development and usage. It promotes explainable, fair, and robust models using a four-step strategy: *explore*, *examine*, *explain* and *expose*. These model-agnostic steps transform complex `ingestibles` (models, data) into interpretable `digestibles`, covering descriptive statistics, performance, local/global explanations, robustness, security, and fairness assessments. Implemented in Python, `Explabox` supports multiple interaction modes, helping developers/testers operationalize explainability, fairness, auditability, and security. The initial release focuses on text data/models, with expansion planned. Code and documentation are available open-source at [https://explabox.readthedocs.io](https://explabox.readthedocs.io/en/stable).
 
 # Statement of need
 
-It is crucial that Machine Learning (ML) development and usage is done in a responsible and transparent manner. High-stakes organizational decisions may significantly impact individuals and society, with potential severe consequences stemming from biases or model errors. This is exemplified by the EU AI Act's regulatory framework, requiring high-risk systems to be properly tested, documented and assessed on their conformity before being applied in practice [@Edwards2022]. Yet, operationalizing transparency (explainable ML) and testing model behavior (fairness, robustness, and security auditing) remains a difficult and laborious task, given the myriad of techniques and their associated learning curves. In response, we have devised a comprehensive four-step analysis strategy&mdash;*explore*, *examine*, *explain*, and *expose*&mdash;ensuring holistic model transparency and testing. The open-source `Explabox` offers these analyses through well-documented, reproducible steps. Data scientists can now access a unified, model-agnostic approach developed and used in a high-stakes context&mdash;the Netherlands National Police&mdash;that is applicable to any text classifier or regressor.
[…]
