+++ paper/paper.md
@@ -38,7 +38,7 @@ bibliography: paper.bib
 
 # Summary
 
-`Explabox` is an open-source toolkit for transparent and responsible machine learning (ML) development and usage. It promotes explainable, fair, and robust models using a four-step strategy: *explore*, *examine*, *explain* and *expose*. These model-agnostic steps transform complex `ingestibles` (models, data) into interpretable `digestibles`, covering descriptive statistics, performance, local/global explanations, robustness, security, and fairness assessments. Implemented in Python, `Explabox` supports multiple interaction modes, helping developers/testers operationalize explainability, fairness, auditability, and security. The initial release focuses on text data/models, with expansion planned. Code and documentation are available open-source at [https://explabox.readthedocs.io](https://explabox.readthedocs.io/en/stable).
+`Explabox` is an open-source toolkit for transparent and responsible machine learning (ML) development and usage. It promotes explainable, fair, and robust models using a four-step strategy: *explore*, *examine*, *explain* and *expose*. These model-agnostic steps transform complex `ingestibles` (models, data) into interpretable `digestibles`, covering descriptive statistics, performance, local/global explanations, and robustness, security and fairness assessments. Implemented in Python, `Explabox` supports multiple interaction modes, helping developers/testers operationalize explainability, fairness, auditability, and security. The initial release focuses on text data/models, with expansion planned. Code and documentation are available open-source at [https://explabox.readthedocs.io](https://explabox.readthedocs.io/en/stable).
 
 # Statement of need
 
@@ -61,15 +61,15 @@ Ingestibles provide a unified model/data import interface, where layers abstract
 
 **Explore** allows data slicing/dicing/sorting, and provides descriptive statistics (dataset sizes, label distributions, and text string/token lengths).
 
-**Examine** shows model performance metrics, summarized in a table or shown graphically, and include include computation/interpretation references. For further analysis, **examine** also supports drilling down into (in)correct predictions.
+**Examine** shows model performance metrics, summarized in a table or shown graphically, and include computation/interpretation references. For further analysis, **examine** also supports drilling down into (in)correct predictions.
 
-**Explain** uses model-agnostic techniques [@Ribeiro2016b] to explain model behavior (*global*) and individual predictions (*local*). It summarizes model-labelled data, through prototypes (`K-Medoids`) or prototypes and criticisms (`MMDCritic` [@Kim2016]), and token distributions (`TokenFrequency`) and informativeness (`TokenInformation`). Local explanations use popular techniques: feature attribution scores (`LIME` [@Ribeiro2016a], `KernelSHAP` [@Lundberg2017]), feature subsets (`Anchors` [@Ribeiro2018]), local rule-based models (`LORE` [@Guidotti2018]), and counterfactual/contrastive explanations (`FoilTrees` [@Waa2018]). Built from generic components separating global and local explanation steps, these methods allow customization and enable scientific advances to be quickly integrated into operational processes (e.g., combine `KernelShap` sampling with `imodels` [@Singh2021] surrogate rules). Example configurations, such as `LIME` with default hyperparameters, easy adoption. **Explain** is provided by subpackage `text_explainability` [@text_explainability], which doubles as a standalone tool.
[…]
