### 2025-04-25 #7904 @MarcelRobeer
@abhishektiwari I have shortened the paper length to <=1000 words. See the latest commit ([5d032e3](https://github.com/MarcelRobeer/explabox/commit/5d032e347de9d0fb86a93f42b56e062de5c2d7c9)) for the changes.
