+++ paper/paper.md
@@ -44,7 +44,7 @@ bibliography: paper.bib
 
 Responsible and transparent Machine Learning (ML) development and usage has become crucial. High-stakes decisions can significantly impact individuals and society, with severe consequences potentially stemming from biases or errors. For example, the EU AI Act requires high-risk systems to be tested, documented, and assessed before practical application [@Edwards2022]. However, operationalizing transparency (explainable ML) and testing model behavior (fairness, robustness, and security auditing) remains difficult due to the myriad techniques and their learning curves. We devised a four-step analysis strategy&mdash;*explore*, *examine*, *explain*, and *expose*&mdash;to ensure holistic model transparency and testing. The open-source `Explabox` offers these analyses through well-documented, reproducible steps. Data scientists gain a unified, model-agnostic approach, developed and used in a high-stakes environment, applicable to any text classifier/regressor.
 
-`Explabox` was developed in an organizational environment where models and data are analyzed repeatedly, and where internal and external stakeholders have varying explanatory needs and preferred formats. Several related tools exist, such as `AIX360` [@Arya2019], `alibi` explain [@Klaise2021], `dalex` [@Baniecki2021], `CheckList` [@Ribeiro2020], and `AIF360` [@Bellamy2018]. However, these tools may be incompatible with recent Python versions (3.8&ndash;3.12), have restricted functionality (testing or explainability focus), lack reproducibility, or lack flexibility communicating results. `Explabox` is an open-source Python toolkit supporting esponsible ML development with minimal workflow disruption for practitioners addressing this gap. Its four-step strategy applies to existing data/models and acts as a hub connecting research and best practices.
+`Explabox` was developed in an organizational environment where models and data are analyzed repeatedly, and where internal and external stakeholders have varying explanatory needs and preferred formats. Several related tools exist, such as `AIX360` [@Arya2019], `alibi` explain [@Klaise2021], `dalex` [@Baniecki2021], `CheckList` [@Ribeiro2020], and `AIF360` [@Bellamy2018]. However, these tools may be incompatible with recent Python versions (3.8&ndash;3.12), have restricted functionality (testing or explainability focus), lack reproducibility, or lack flexibility communicating results. `Explabox` is an open-source Python toolkit supporting responsible ML development with minimal workflow disruption for practitioners addressing this gap. Its four-step strategy applies to existing data/models and acts as a hub connecting research and best practices.
 
 # Explore, Examine, Explain & Expose your ML models
 
