### 2025-05-09 #7904 @MarcelRobeer
@abhishektiwari my apologies for the delayed reply, I am in the process of finishing up my PhD dissertation so I was heavily focused on that.

I have no specific suggestions for specific reviewers. However, I can affirm that the five most similar historical JOSS papers as found by `editorialbot` are similar topic-wise, so reviewers from these papers would potentially also be a relevant fit for this paper as well.
