### 2025-05-09 #7904 @axsaucedo
Thanks for the mention @abhishektiwari - to be honest this is the first time I come across this JOSS repo, so not sure i have much context. Looking at the PDF overall it looks like a comprehensive analysis of some explainability libraries, however if I am to provide a very high level opinion, it seems to me that this package is neither widely adopted nor very actively maintained (looking at issues / contribution graph / PR / etc), so depending on the criteria this may or may not have to be considered. I maintain a list of production ML tools, one section on explainability which has some of the tools mentioned together with others which may be more relevant (or at least additionally relevant): https://github.com/EthicalML/awesome-production-machine-learning/?tab=readme-ov-file#explainability-and-fairness. Hope these initial thoughts are helpful - feel free to ignore otherwise.
