### 2025-05-09 #7904 @MarcelRobeer
@abhishektiwari (and reviewers), please note that — as stated in the paper — the Explabox does not only consist of the Explabox repository itself, but its main functionality comes from two subpackages https://github.com/MarcelRobeer/text_explainability and https://github.com/MarcelRobeer/text_sensitivity, that also double as standalone tools. It was an architectural choice with the software development team to split them up into subpackages for the overall framework and data handling (`explabox`), explainability (`text_explainability`; 5771 LOC / 86K pip downloads) and fairness/robustness/security sensitivity analyses (`text_sensitivity`; 6290 LOC / 56K pip downloads). The latter two packages contain full modular re-implementations for model-agnostic explainability and sensitivity analysis, respectively, including traceability. The idea behind splitting it up was to make it future-proof for other data modalities, where those will have corresponding packages.

All packages are ports that originated in a secured git environment, but we decided to share them open-source such that the community can enjoy them beyond just internal usage. I decided it would be best to bundle all of them into a single publication instead of three separate ones, which ended up being the current submission.

Hopefully this clarifies things a bit. Thank you again for your consideration of the software and paper.
