### 2025-05-09 #7904 @abhishektiwari
@axsaucedo Thanks for getting back and providing feedback. JOSS follows a checklist-driven reviews using GitHub issues. This is pre-review Github issue. Once we start review using a new Github issue, reviewers will have opportunity to provide detailed feedback on the software including if software represents substantial scholarly effort and whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain, quality of documentation, software functionality, etc. See an example checklist for an ongoing review [here](https://github.com/openjournals/joss-reviews/issues/8038#issuecomment-2797107623).

Popularity of software project in terms of no. of issues / contribution graph / PR / stars / etc is not a criteria for publishing in JOSS. Re: ongoing maintenance as @MarcelRobeer mentioned other two packages covered by paper have been under active development during 6 months.

I hope it makes sense. Let me know if you will be open to review this submission.
