### 2025-05-10 #7904 @axsaucedo
Sounds good, I like this JOSS initaitive, hadn't heard about it before - at this stage unfortunately I've committed to too many other things so will have to pass on this one, but let me know for other future ones in case possible in the near future, particularly relevant to any category from the list provided above
