### 2025-05-13 #7904 @abhishektiwari
👋 @Naviden @klainfo @JHoelli @jklaise, would any of you be willing to review this software paper submission for JOSS? The submission under consideration is "The Explabox: Model-Agnostic Machine Learning Transparency & Analysis".

Based on your experience in Explainable AI (XAI) space, we think you might be able to provide a great review of this submission. If not, could you recommend any potential reviewers?

Thanks
Abhi
