### 2025-05-18 #7904 @abhishektiwari
@rachitanaik-lyft would you be willing to review this software paper submission for JOSS? The submission under consideration is "The Explabox: Model-Agnostic Machine Learning Transparency & Analysis".

Based on your expertise/interest in AI/ML space, I think you might be able to provide a great review of this submission.

Thanks
Abhi
