### 2025-06-07 #8253 @abhishektiwari
Hi @hbaniecki, @JHoelli, @rachitanaik-lyft thanks for agreeing to review this submission. Please create your review checklist by running,

To suggest improvement, please create a new issue in the [target repository](https://github.com/MarcelRobeer/explabox) and link to those issues (especially acceptance-blockers) by leaving comments in this review thread.
