+++ README.md
@@ -49,6 +49,8 @@ The `explabox` is distributed on [PyPI](https://pypi.org/project/explabox/). To 
 ...                splits={'train': 'drugsComTrain.tsv', 'test': 'drugsComTest.tsv'})
 ```
 
+The example dataset and model used can be easily imported using demo package `explabox-demo-drugreview`.
+
 Then `.explore`, `.examine`, `.expose` and `.explain` your model:
 ```python
 >>> # Explore the descriptive statistics for each split
@@ -115,11 +117,11 @@ The `explabox` consists of three layers:
 
 <a name="example-usage"/></a>
 ## Example usage
-The [example usage guide](EXAMPLE_USAGE.md) showcases the `explabox` for a black-box model performing multi-class classification of the [UCI Drug Reviews](https://archive.ics.uci.edu/ml/datasets/Drug+Review+Dataset+%28Drugs.com%29#) dataset.
+The [example usage guide](EXAMPLE_USAGE.md) showcases the `explabox` for a black-box model performing multi-class classification of the [UCI Drug Reviews](https://archive.ics.uci.edu/dataset/461/drug+review+dataset+druglib+com) dataset.
 
 Without requiring any local installations, the notebook is provided on [![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/14lXvXV01DaSruSAhD1RLbILRl2mPQ4nS?usp=sharing).
 
-If you want to follow along on your own device, simply `pip install explabox-demo-drugreview` and run the lines in the [Jupyter notebook](https://git.science.uu.nl/m.j.robeer/explabox-demo-drugreview/-/blob/main/explabox_demo_drugreview.ipynb) we have prepared for you!
+If you want to follow along on your own device, simply `pip install explabox-demo-drugreview` and run the lines in the [Jupyter notebook](https://colab.research.google.com/drive/14lXvXV01DaSruSAhD1RLbILRl2mPQ4nS?usp=sharing) we have prepared for you!
 
 <a name="releases"/></a>
 ## Releases
