### 2025-06-30 e89da05
Add bandit skip
Signed-off-by: Marcel Robeer <marcel_robeer@hotmail.com>
