### 2025-07-04 #8253 @abhishektiwari
@MarcelRobeer - Did we provide response/fixes on Github ticket opened by @JHoelli?

@hbaniecki - please feel free to open [Github issue on Explabox repository](https://github.com/MarcelRobeer/explabox/issues) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below.
