### 2025-07-05 #8253 @MarcelRobeer
> [@MarcelRobeer](https://github.com/MarcelRobeer) - Did we provide response/fixes on Github ticket opened by [@JHoelli](https://github.com/JHoelli)?
>
> [@hbaniecki](https://github.com/hbaniecki) - please feel free to open [Github issue on Explabox repository](https://github.com/MarcelRobeer/explabox/issues) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below.

@abhishektiwari I have responded to both issues by @JHoelli - fixed one by updating a URL to a dataset and am aiwiting her confirmation that the second issue is not a bug but due to missing a step in the example usage.

@abhishektiwari For the issues related to the software paper (minor revisions), am I expected to address them immediately or first wait untill all reviewers have provided feedback?
