### 2025-07-14 d3cc79d
Updated overview in README.md (Fixed #14)
