### 2025-07-16 #8253 @hbaniecki
I have the following concerns about this submission:

1. **Substantial scholarly effort is (currently) not evident.** Since 2022, the software has gathered only 17K downloads; there seems to be no evident user base (a few issues by the paper's authors, 17 GitHub stars, no blog posts and/or promotion of the software?), and no clear applications (e.g. no citations, insightful use cases). With similar software existing already, {Explabox} is not likely to be cited unless the software/paper clearly highlights its original purpose, e.g. an original subset of methods implemented, accessible visualisation, better API, unique audience.
2. **Statement of need is unclear.** I have a hard time understanding the motivation for developing this software. Currently, the paper's title and description are too generic.
    - Should 'text' be highlighted in the title? 'Model-agnostic' is misleading since the software only considers models for the text modality. 'Transparency' is vague and not even defined in the paper.
    - The first paragraph (L18-28) focuses on 'high-stakes decisions' and 'high-risk systems' without describing how this software relates to them. It is unclear whether proposing another package explaining text classification solves a concrete problem. Describing a clear use case for this software would help (e.g. from the police/law domain).
    - The described strategy 'explore, examine, explain and expose' resembles the notion of 'explore, explain, and examine' from https://doi.org/10.1201/9780429027192, which could be referenced.
3. **State of the field is too general.** There is only a vague description of how this software compares to other similar packages. The second paragraph (L29-38) mentions that other tools exist (L31-33) but (L33-35):
    - 'these may be incompatible with recent Python versions (3.8–3.12)' Which ones? Is it even a crucial research contribution?
    - 'have restricted functionality (testing or explainability focus)' Which functionality?
    - 'lack reproducibility' How?
    - 'or lack flexibility communicating results.' What do you mean by that?
The remaining two sentences (L35-38) also seem vague, mentioning 'minimal workflow disruption' and 'best practices'. Relating {Explabox} to other software could partially alleviate concerns mentioned in 1. & 2., i.e. show the added value of {Explabox} over the already available solutions, motivate the need for such a new package, highlight the target audience etc.

I am open to discussion; perhaps the software paper can be improved to clearly state the motivation and effort.

As a sidenote, a software with a similar profile was published in JOSS 1.5 years ago and has gained zero citations to date, excluding one by the same authors. My concerns discussed in that particular case https://github.com/openjournals/joss-reviews/issues/5873#issuecomment-1761368406 remain relevant, being applicable to this submission as well.

**Relevant references**
Alber et al. iNNvestigate Neural Networks! JMLR 2019
Kokhlikyan et al. Captum: A unified and generic model interpretability library for PyTorch. arXiv:2009.07896 2020
Li et al. InterpretDL: Explaining Deep Models in PaddlePaddle. JMLR 2022
Sarti et al. Inseq: An Interpretability Toolkit for Sequence Generation Models. ACL 2023
