### 2025-07-17 #8253 @abhishektiwari
Thanks @hbaniecki. Generally JOSS review criteria does not consider downloads, size of user base, popularity in terms of stars or clones etc. However the software is sufficiently useful that it is likely to be cited or have been already cited is important consideration. I will let @MarcelRobeer respond to your feedback.
