### 2025-07-21 MarcelRobeer/explabox#15 @hbaniecki
Error when executing demo from readme
Hi,
1. Is the demo working correctly? Executing `box.expose.compare_metrics(split='test', perturbation='add_typos')` runs into `AttributeError: 'Exposer' object has no attribute 'compare_metrics'`, while using `box.expose.compare_metric(splits='test', perturbation='add_typos')` runs into `TypeError: bad operand type for abs(): 'str'`. I tested it on both the PyPI and GitHub package versions.
2. I don't know the purpose of writing ">>> " in code examples, which seems to make copying the code harder. I found that another example given at https://explabox.readthedocs.io/en/latest/example-usage.html does not contain these symbols.

CC: https://github.com/openjournals/joss-reviews/issues/8253
