### 2025-08-07 74071e4
Update EXAMPLE_USAGE.md
