### 2025-08-29 bf9ebba
Added new source to bib
