+++ paper/paper.bib
@@ -205,3 +205,11 @@
     year={2022},
     url={https://www.adalovelaceinstitute.org/wp-content/uploads/2022/04/Expert-explainer-The-EU-AI-Act-11-April-2022.pdf}
 }
+@book{Biecek2021,
+    title = {{Explanatory Model Analysis: Explore,  Explain and Examine Predictive Models}},
+    url = {http://dx.doi.org/10.1201/9780429027192},
+    DOI = {10.1201/9780429027192},
+    publisher = {Chapman and Hall/CRC},
+    author = {Biecek, Przemyslaw and Burzykowski, Tomasz},
+    year = {2021},
+}
