### 2025-08-29 bdc64c7
New paper version incorporating reviewer feedback
