+++ paper/paper.md
@@ -1,13 +1,13 @@
 ---
-title: 'Explabox: Model-Agnostic Machine Learning Transparency & Analysis'
+title: 'Explabox: A Python Toolkit for Standardized Auditing and Explanation of Text Models'
 tags:
     - Python
+    - AI auditing
     - explainable AI (XAI)
     - interpretability
     - fairness
     - robustness
     - AI safety
-    - AI auditing
 authors:
   - name: Marcel Robeer
     orcid: 0000-0002-6430-9774
@@ -32,26 +32,28 @@ affiliations:
   - name: School of Law, Utrecht University, The Netherlands
     index: 3
     ror: 04pp8hn57
-date: 13 March 2025
+date: 29 August 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-`Explabox` is an open-source toolkit for transparent and responsible machine learning (ML) development and usage. It promotes explainable, fair, and robust models using a four-step strategy: *explore*, *examine*, *explain* and *expose*. These model-agnostic steps transform complex `ingestibles` (models, data) into interpretable `digestibles`, covering descriptive statistics, performance, local/global explanations, and robustness, security and fairness assessments. Implemented in Python, `Explabox` supports multiple interaction modes, helping developers/testers operationalize explainability, fairness, auditability, and security. The initial release focuses on text data/models, with expansion planned. Code and documentation are available open-source at [https://explabox.readthedocs.io](https://explabox.readthedocs.io/en/stable).
+Developed to meet the practical machine learning (ML) auditing requirements of the Netherlands National Police, `Explabox` is an open-source Python toolkit that implements a standardized four-step analysis workflow: *explore*, *examine*, *explain* and *expose*. The framework transforms models and data (*ingestibles*) into interpretable reports and visualizations (*digestibles*), covering everything from data statistics and performance metrics to local/global explanations, and sensitivity testing for fairness, robustness and security. Designed for developers, testers, and auditors, `Explabox` operationalizes the entire audit lifecycle in a reproducible manner. The initial release is focused on text classification and regression models, with plans for future expansion. Code and documentation are available open-source at [https://explabox.readthedocs.io](https://explabox.readthedocs.io/en/stable).
 
 # Statement of need
 
-Responsible and transparent Machine Learning (ML) development and usage has become crucial. High-stakes decisions can significantly impact individuals and society, with severe consequences potentially stemming from biases or errors. For example, the EU AI Act requires high-risk systems to be tested, documented, and assessed before practical application [@Edwards2022]. However, operationalizing transparency (explainable ML) and testing model behavior (fairness, robustness, and security auditing) remains difficult due to the myriad techniques and their learning curves. We devised a four-step analysis strategy&mdash;*explore*, *examine*, *explain*, and *expose*&mdash;to ensure holistic model transparency and testing. The open-source `Explabox` offers these analyses through well-documented, reproducible steps. Data scientists gain a unified, model-agnostic approach, developed and used in a high-stakes environment, applicable to any text classifier/regressor.
[…]
