### 2025-08-29 #8253 @MarcelRobeer
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08253/joss.08253/10.21105.joss.08253.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.08253/joss.08253/10.21105.joss.08253.pdf) 📄 👈

In the latest paper version, I have incorporated the reviewer's concerns:

1. **Updated motivation/scope**: the paper now centers on the original intended use case (auditing text models at the Netherlands National Police), with an updated _title_ and _statement of need_ that clearly define this specific, high-stakes origin and contribution. (@hbaniecki)
2. **Strengthened state of the field**: vague claims were replaced with a specific comparative analysis, including an early differentiation from similar libraries like `OmniXAI` and `Dalex` to immediately clarify the software's unique workflow-centric niche. (@hbaniecki, @JHoelli)
3. **Provided concrete justifications**: the revision now includes specific technical details on how `Explabox` addresses key operational challenges, such as ensuring reproducibility via centralized seed/parameter tracking and providing flexible reporting for diverse stakeholders through the _digestible_ object system. (@hbaniecki)
