### 2025-08-30 #8253 @hbaniecki
Hi, the paper reads better now. There is a typo in "OmniXAI (?)". I finished my checklist and can recommend acceptance.
