+++ paper/paper.bib
@@ -38,11 +38,11 @@
     year = {2018}
 }
 @article{Guidotti2018,
-    title={{Local Rule-Based Explanations of Black Box Decision Systems}},
-    author={Guidotti, Riccardo and Monreale, Anna and Ruggieri, Salvatore and Pedreschi, Dino and Turini, Franco and Giannotti, Fosca},
-    journal={arXiv preprint},
-    doi={arXiv.1805.10820},
-    year={2018}
+    title = {{Local Rule-Based Explanations of Black Box Decision Systems}},
+    author = {Guidotti, Riccardo and Monreale, Anna and Ruggieri, Salvatore and Pedreschi, Dino and Turini, Franco and Giannotti, Fosca},
+    journal = {arXiv preprint},
+    doi = {arXiv.1805.10820},
+    year = {2018}
 }
 @inproceedings{Kim2016,
     author = {Kim, Been and Khanna, Rajiv and Koyejo, Oluwasanmi},
@@ -51,25 +51,25 @@
     year = {2016}
 }
 @article{Mehrabi2021,
-    title={{A Survey on Bias and Fairness in Machine Learning}},
-    author={Mehrabi, Ninareh and Morstatter, Fred and Saxena, Nripsuta and Lerman, Kristina and Galstyan, Aram},
-    journal={ACM Computing Surveys (CSUR)},
-    volume={54},
-    number={6},
-    pages={1--35},
-    year={2021},
-    doi={10.1145/3457607},
-    publisher={ACM New York, NY, USA}
+    title = {{A Survey on Bias and Fairness in Machine Learning}},
+    author = {Mehrabi, Ninareh and Morstatter, Fred and Saxena, Nripsuta and Lerman, Kristina and Galstyan, Aram},
+    journal = {ACM Computing Surveys (CSUR)},
+    volume = {54},
+    number = {6},
+    pages = {1--35},
+    year = {2021},
+    doi = {10.1145/3457607},
+    publisher = {ACM New York, NY, USA}
 }
 @misc{instancelib,
-   author       = {Bron, Michiel P.},
-   title        = {{Python Package instancelib}},
-   month        = sep,
-   year         = 2023,
-   publisher    = {Zenodo},
-   version      = {0.5.0},
-   doi          = {10.5281/zenodo.8308017},
-   url          = {https://doi.org/10.5281/zenodo.8308017}
+   author = {Bron, Michiel P.},
+   title = {{Python Package instancelib}},
+   month = sep,
+   year = 2023,
+   publisher = {Zenodo},
+   version = {0.5.0},
+   doi = {10.5281/zenodo.8308017},
+   url = {https://doi.org/10.5281/zenodo.8308017}
 }
 @misc{text_explainability,
     title = {Python package text\_explainability},
@@ -136,15 +136,15 @@
     url = {https://github.com/joke2k/faker}
 }
 @article{scikit-learn,
-    title={{Scikit-learn: Machine Learning in {P}ython}},
-    author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+    title = {{Scikit-learn: Machine Learning in {P}ython}},
+    author = {Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
          and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
          and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
          Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
-    journal={Journal of Machine Learning Research},
-    volume={12},
-    pages={2825--2830},
-    year={2011}
[…]
