### 2025-09-04 #8253 @abhishektiwari
Hi @JHoelli, just wanted to circle back on the revisions @MarcelRobeer completed and if you've had a chance to look at the updates.

> 👋 [@JHoelli](https://github.com/JHoelli) - [@MarcelRobeer](https://github.com/MarcelRobeer) has completed revisions to both the paper and software based on the feedback provided. Could you please review the updates when you have a chance and let us know whether the changes adequately address your feedback?
