### 2025-09-19 #8253 @JHoelli
Hi @abhishektiwari , sorry for the delayed response. I’m currently quite busy, but I’ll get it done next week.

One question: if I remember correctly, all of the code was written in the __init__.py file, which makes it a bit difficult to navigate and isn’t really in line with software engineering best practices. @MarcelRobeer , was there a specific reason for structuring it that way?

And @abhishektiwari , do you think this is something that should be addressed ( ist not Something that influences the functionality)?
