### 2025-09-21 #8253 @abhishektiwari
Hi @JHoelli  I looked at codebase, I can see unusual code file structure here that said I don't think this is a blocker.

@MarcelRobeer two comments

Firstly, issue mentioned above can be very easily refactored without introducing any breaking changes.

For example take [this file](https://github.com/MarcelRobeer/explabox/blob/main/explabox/digestibles/__init__.py). You can move code to a new file lets say `classes.py` and then import in `__init__.py`

```
from .classes import Performance, Dataset
__all__ = ["Performance", "Dataset"]
```

Secondly, I see a different copyright notification in some of the code files. Can you please clarify and fix?

```
# Copyright (C) 2021-2022 National Police Lab AI (NPAI).
```
