### 2025-09-24 #8253 @JHoelli
Hi, I also finished my review list. The provided fixes work, and the paper has improved quite a bit.
As my last point regarding the software engineering is not a blocker, I can also recommend acceptance.

However, I would strongly encourage improving the code structure for better usability. (A good code structure is often more helpful than the documentation itself, and in this case, it took me quite some time to find the methods I was looking for. Moreover, following best practices can help attract project collaborators who may want to contribute their own methods or further advance the development of the library.)
