### 2025-09-30 ccc6ab9
Code restructure and copyright notice
