### 2025-09-30 5e04ada
Code restructure
