+++ explabox/explain/text/explainer.py
@@ -1,0 +1,357 @@
+# Copyright (c) 2022 Marcel Robeer for National Police Lab AI (NPAI).
+#
+# This program is free software; you can redistribute it and/or modify it under the terms of the GNU Lesser General
+# Public License (LGPL) as published by the Free Software Foundation; either version 3 (LGPLv3) of the License, or (at
+# your option) any later version. You may not use this file except in compliance with the license. You may obtain a copy
+# of the license at:
+#
+#     https://www.gnu.org/licenses/lgpl-3.0.en.html
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more
+# details.
+
+"""Main Explainer class."""
+
+import warnings
+from multiprocessing.sharedctypes import Value
+from typing import List, Optional, Union
+
+from genbase import Readable, translate_list
+from instancelib import AbstractClassifier, Environment
+from text_explainability.data.embedding import Embedder, TfidfVectorizer
+from text_explainability.generation.return_types import FeatureList, Instances
+
+from ...ingestibles import Ingestible
+from ...mixins import IngestiblesMixin
+from ...ui.notebook import restyle
+from ...utils import MultipleReturn
+
+
+class Explainer(Readable, IngestiblesMixin):
+    def __init__(
+        self,
+        data: Optional[Environment] = None,
+        model: Optional[AbstractClassifier] = None,
+        ingestibles: Optional[Ingestible] = None,
+        **kwargs,
+    ):
+        """The Explainer creates explanations corresponding to a model and dataset (with ground-truth labels).
+
+        With the Explainer you can use explainble AI (XAI) methods for explaining the whole dataset (global), model
+        behavior on the dataset (global), and specific predictions/decisions (local).
+
+        The Explainer requires 'data' and 'model' defined. It is included in the Explabox under the `.explain` property.
+
+        Examples:
+            Construct the explainer:
+
+            >>> from explabox.explain import Explainer
+            >>> explainer = Explainer(data=data, model=model)
+
+            Get a local explanation with LIME (https://github.com/marcotcr/lime) and kernelSHAP
+            (https://github.com/slundberg/shap):
+
+            >>> explainer.explain_prediction('I love this so much!', methods=['lime', 'kernel_shap'])
+
+            See the top-25 tokens for predicted classifier labels on the test set:
+
+            >>> explainer.token_frequency(k=25, explain_model=True, splits='test')
+
+            Select the top-5 prototypical examples in the train set:
+
+            >>> explainer.prototypes(n=5, splits='train')
+
+        Args:
+            data (Optional[Environment], optional): Data for ingestibles. Defaults to None.
+            model (Optional[AbstractClassifier], optional): Model for ingestibles. Defaults to None.
+            ingestibles (Optional[Ingestible], optional): Ingestible. Defaults to None.
+        """
+        if ingestibles is None:
+            ingestibles = Ingestible(data=data, model=model)
+        self.ingestibles = ingestibles
+        self.check_requirements(["data", "model"])
+
+    @restyle
+    def explain_prediction(
+        self,
+        sample: Union[int, str],
+        *args,
+        methods: Union[str, List[str]] = ["lime"],
+        **kwargs,
+    ) -> Optional[MultipleReturn]:
+        """Explain specific sample locally.
+
+        Args:
+            sample: Identifier of sample in dataset (int) or input (str).
+            methods: List of methods to get explanations from. Choose from 'lime', 'shap', 'baylime', 'tree', 'rules', 'foil_tree'.
[…]
