+++ explabox/explain/text/explainer.py
@@ -14,7 +14,6 @@
 """Main Explainer class."""
 
 import warnings
-from multiprocessing.sharedctypes import Value
 from typing import List, Optional, Union
 
 from genbase import Readable, translate_list
@@ -84,7 +83,8 @@ class Explainer(Readable, IngestiblesMixin):
 
         Args:
             sample: Identifier of sample in dataset (int) or input (str).
-            methods: List of methods to get explanations from. Choose from 'lime', 'shap', 'baylime', 'tree', 'rules', 'foil_tree'.
+            methods: List of methods to get explanations from. Choose from 'lime', 'shap', 'baylime', 
+                'tree', 'rules', 'foil_tree'.
             *args: Positional arguments passed to local explanation technique.
             **kwargs: Keyword arguments passed to local explanation technique.
 
