+++ explabox/ingestibles/__init__.py
@@ -18,5 +18,5 @@ from .data import import_data, rename_labels, train_test_split
 from .ingestible import Ingestible
 from .model import import_model
 
-__all__ = ["Ingestible",
-           "import_data", "import_model", "rename_labels", "train_test_split"]
+__all__ = ["Ingestible", "import_data", "import_model", "rename_labels", "train_test_split"]
+d
