+++ .github/pull_request_template.md
@@ -8,8 +8,8 @@
 
 [//]: # (Have you performed all steps in accordance with the contribution guide at https://explabox.readthedocs.io/en/latest/contributing.html?)
 ## Delivery checklist
-[ ] The contribution contains one logical unit
-[ ] I have performed all quality checks
-[ ] (_in case of updated software_) I have included the corresponding tests for the software I contributed
-[ ] (_in case of updated software_) My software contribution is fully documented
-[ ] I have updated `CHANGELOG.md` under the `[Unreleased]` section
+- [ ] The contribution contains one logical unit
+- [ ] I have performed all quality checks
+- [ ] (_in case of updated software_) I have included the corresponding tests for the software I contributed
+- [ ] (_in case of updated software_) My software contribution is fully documented
+- [ ] I have updated `CHANGELOG.md` under the `[Unreleased]` section
