### 2025-10-02 5434645
Update README.md to match paper
