### 2025-10-02 d35d7ff
Updated MANIFEST.in
