+++ CITATION.cff
@@ -21,7 +21,6 @@ authors:
   - given-names: Floris
     family-names: Bex
     orcid: 'https://orcid.org/0000-0002-5699-9656'
-doi: 10.48550/arXiv.2411.15257
 repository-code: 'https://github.com/MarcelRobeer/explabox'
 url: 'https://explabox.readthedocs.io'
 license: LGPL-3.0
