### 2025-10-03 d5884a6
Exposed return types for testing
