+++ explabox/_version.py
@@ -13,5 +13,5 @@
 
 """Version information."""
 
-__version_info__ = (1, 0, 2)
+__version_info__ = (1, 0, 3)
 __version__ = ".".join(str(v) for v in __version_info__)
