### 2025-10-05 9d84942
Made the paper a bit more verbose
