+++ paper/paper.md
@@ -42,7 +42,7 @@ Developed to meet the practical machine learning (ML) auditing requirements of t
 
 # Statement of need
 
-In high-stakes environments like law enforcement, machine learning (ML) models are subject to intense scrutiny and must comply with emerging regulations like the EU AI Act [@Edwards2022]. `Explabox` was developed to address the operational challenges of ML auditing at the Netherlands National Police, where models for text classification/regression require standardized, reproducible, and holistic evaluation to satisfy diverse stakeholders&mdash;from developers and internal auditors to legal and ethical oversight bodies. Existing tools, while powerful, were often fragmented, focusing on a single aspect of analysis (e.g., only explainability or testing) and lacking a unified framework for conducting a complete audit from data exploration to final reporting.
+In high-stakes environments like law enforcement, machine learning (ML) models are subject to intense scrutiny and must comply with emerging regulations like the EU AI Act [@Edwards2022]. `Explabox` was developed to address the operational challenges of ML auditing at the Netherlands National Police, where models for text classification and regression require standardized, reproducible, and holistic evaluation to satisfy diverse stakeholders&mdash;from developers and internal auditors to legal and ethical oversight bodies. Existing tools, while powerful, were often fragmented, focusing on a single aspect of analysis (e.g., only explainability or testing) and lacking a unified framework for conducting a complete audit from data exploration to final reporting.
 
 To solve this workflow problem, we developed Explabox around a four-step analysis strategy&mdash;*explore*, *examine*, *explain* and *expose*&mdash;inspired by similar conceptualizations of the analytical process [@Biecek2021]. While comprehensive libraries like `OmniXAI` [@Yang2022] offer a broad, multi-modal collection of explainers and `dalex` [@Baniecki2021] provides a mature, research-driven framework for model exploration, `Explabox` was developed to fill a specific operational gap. Practitioners seeking to conduct a full audit in a model-agnostic manner often have to combine multiple, highly-specialized libraries, such as `AIF360` [@Bellamy2018] for fairness metrics, `alibi explain` [@Klaise2021] or `AIX360` [@Arya2019] for local explanations, and `CheckList` [@Ribeiro2020] for behavioral testing.
 
@@ -53,7 +53,7 @@ This fragmentation introduces significant challenges, particularly regarding *re
 `Explabox` transforms opaque *ingestibles* into transparent *digestibles* through four types of *analyses* to enhance explainability and aid fairness, robustness, and security audits.
 
 ## Ingestibles
-Ingestibles provide a unified model/data import interface, where layers abstract away access (\autoref{fig:layers}) to allow optimized processing. `Explabox` uses `instancelib` [@instancelib] for fast model/data encapsulation. The model can be any Python `Callable` containing a regression or (binary and multi-class) classification model.  While this interface is model-agnostic, the current release provides data handling and analysis modules optimized specifically for text-based tasks. `scikit-learn` or `onnx` models (e.g., PyTorch, TensorFlow/Keras) import directly with optimizations and automatic input/output interpretation. Data can be automatically downloaded, extracted and loaded. Data inputs include `NumPy`/`Pandas`/`huggingface`, raw files (e.g., HDF5, CSV or TSV), and (compressed) file folders. Data can be subdivided into named splits (e.g., train-test-validation), and instance vectors and tokens can be precomputed (and optionally saved) for fast inferencing.
[…]
