### 2025-10-05 c2bd9ea
Minor paper fixes
