### 2025-10-05 #8253 @MarcelRobeer
> [@MarcelRobeer](https://github.com/MarcelRobeer)
>
> I suggest you review all occurrences of and consider changing the slashes to commas, or, and as relevant and appropriate.
>
> Line 21: consider changing "classification/regression" to "classification or regression"? Line 56: "onnx" should be "ONNX", consider changing "PyTorch, TensorFlow/Keras" to "PyTorch, TensorFlow, or Keras" Line 58: change to "downloaded, extracted, and loaded" Line 58 - instead of "huggingface" use "Hugging Face". Also, consider changing "NumPy/Pandas/huggingface" to "NumPy, Pandas, or Hugging Face". Line 59: consider changing "and (compressed) file folders" Line 60: consider changing "and instance vectors and tokens can be precomputed and optionally saved for fast inferencing" Line 65: "slicing/dicing/sorting" - Consider changing to "slicing, dicing, or sorting" Line 68: "include computation/interpretation references" change to "includes computation or interpretation references" Line 76: consider changing "counterfactual/contrastive explanations" to "counterfactual or contrastive explanations" Line 12, 83: consider changing "local/global" to "local or global" Line 89 - "multi-language data generation" - While not technically wrong, "multilingual" or "multi-lingual" is more standard and grammatically preferred
>
> Once done. Let me know I will review the paper again.

@abhishektiwari I have updated the affiliations on Zenodo and changed the `paper.md` file to incorporate your suggested changes.
