+++ paper/paper.md
@@ -63,7 +63,7 @@ Ingestibles provide a unified model/data import interface, where layers abstract
 
 **Explore** allows data slicing, dicing and sorting, and provides descriptive statistics (dataset sizes, label distributions, and text string/token lengths).
 
-**Examine** shows model performance metrics, summarized in a table or shown graphically with computation and interpretation references. For further analysis, **examine** also supports drilling down into (in)correct predictions.
+**Examine** shows model performance metrics, summarized in a table or shown graphically, with computation and interpretation references. For further analysis, **examine** also supports drilling down into (in)correct predictions.
 
 **Explain** uses model-agnostic techniques [@Ribeiro2016b] to explain model behavior (*global*) and individual predictions (*local*). It summarizes model-labelled data, through prototypes (`K-Medoids`) or prototypes and criticisms (`MMDCritic` [@Kim2016]), and token distributions (`TokenFrequency`) and informativeness (`TokenInformation`). Local explanations use popular techniques: feature attribution scores (`LIME` [@Ribeiro2016a], `KernelSHAP` [@Lundberg2017]), feature subsets (`Anchors` [@Ribeiro2018]), local rule-based models (`LORE` [@Guidotti2018]), and counterfactual or contrastive explanations (`FoilTrees` [@Waa2018]). Built from generic components separating global and local explanation steps, these methods allow customization and enable scientific advances to be quickly integrated into operational processes (e.g., combine `KernelShap` sampling with `imodels` [@Singh2021] surrogate rules). Example configurations, such as `LIME` with default hyperparameters, ease adoption. **Explain** is provided by subpackage `text_explainability` [@text_explainability], which doubles as a standalone tool.
 
