### 2025-10-06 #8253 @abhishektiwari
@MarcelRobeer can you please check following instances,

Line 12: "local/global explanations"
Line 51: "model/data import interface"
Line 53: "model/data encapsulation"
Line 97: "Jupyter/web UI", also capitalise "Web UI"

Once you confirm, I can move on to make recommendation.
