### 2025-10-06 #8253 @abhishektiwari
@MarcelRobeer  -

For style consistency can we ensure Oxford commas are used when listing  three or more items. Some places you use but then others you don't.

LINE 13: "sensitivity testing for fairness, robustness and security." On same line used correctly for "developers, testers, and auditors"
LINE 10, 28, 64: "explore, examine, explain and expose."
LINE 59: "HDF5, CSV or TSV"
LINE 96: "creators, auditors, applicants, end-users or clients"

LINE 71-73: "It summarizes model-labelled data, through prototypes (K-Medoids) or prototypes and criticisms (MMDCritic (Kim et al., 2016)), and token distributions (TokenFrequency) and informativeness (TokenInformation)." can be better reframed for clarity, e.g.,

"It summarizes model-labelled data through prototypes (K-Medoids), prototypes with criticisms (MMDCritic (Kim et al., 2016)), token distributions (TokenFrequency), and token informativeness (TokenInformation)."

OR
"It summarizes model-labelled data through prototypes (K-Medoids) or prototypes and criticisms (MMDCritic (Kim et al., 2016)), as well as through token distributions (TokenFrequency) and token informativeness (TokenInformation)."

OR
"It summarizes model-labelled data using several techniques: prototypes (K-Medoids), prototypes with criticisms (MMDCritic (Kim et al., 2016)), token distributions (TokenFrequency), and token informativeness (TokenInformation)."
