### 2025-10-07 19bac63
Consistent usage of Oxford commas
