### 2025-10-07 #8253 @abhishektiwari
Arxiv doi starts with 10.48550. E.g., 10.48550/arXiv.1805.10820
