+++ paper/paper.bib
@@ -41,7 +41,7 @@
     title = {{Local Rule-Based Explanations of Black Box Decision Systems}},
     author = {Guidotti, Riccardo and Monreale, Anna and Ruggieri, Salvatore and Pedreschi, Dino and Turini, Franco and Giannotti, Fosca},
     journal = {arXiv preprint},
-    doi = {arXiv.1805.10820},
+    doi = {10.48550/arXiv.1805.10820},
     year = {2018}
 }
 @inproceedings{Kim2016,
@@ -106,7 +106,6 @@
     pdf = {http://proceedings.mlr.press/v97/agarwal19d/agarwal19d.pdf},
 }
 @inproceedings{Ribeiro2016b,
-    arxivId = {1606.05386},
     author = {Ribeiro, Marco Tulio and Singh, Sameer and Guestrin, Carlos},
     booktitle = {2016 ICML Workshop on Human Interpretability in Machine Learning (WHI 2016)},
     doi = {10.1145/2858036.2858529},
@@ -165,7 +164,7 @@
     month = sep,
     year = {2019},
     journal = {arXiv preprint},
-    doi = {arXiv.1909.03012}
+    doi = {10.48550/arXiv.1909.03012}
 }
 @article{Bellamy2018,
     title = {{AI Fairness 360: An Extensible Toolkit for Detecting, Understanding, and Mitigating Unwanted Algorithmic Bias}},
@@ -178,7 +177,7 @@
     month = oct,
     year = {2018},
     journal = {arXiv preprint},
-    doi = {arXiv.1810.01943}
+    doi = {10.48550/arXiv.1810.01943}
 }
 @article{Klaise2021,
     author = {Janis Klaise and Arnaud Van Looveren and Giovanni Vacanti and Alexandru Coca},
@@ -217,7 +216,6 @@
     title = {{OmniXAI: A Library for Explainable AI}},
     author = {Wenzhuo Yang and Hung Le and Silvio Savarese and Steven Hoi},
     year = {2022},
-    url = {https://arxiv.org/abs/2206.01612},
-    archivePrefix = {arXiv},
-    eprint = {206.01612},
+    journal = {arXiv preprint},
+    doi = {10.48550/arXiv.2206.01612},
 }
