### 2025-10-07 #8253 @MarcelRobeer
@abhishektiwari DOIs have been fixed in the [latest commit](https://github.com/MarcelRobeer/explabox/commit/135421e3db6e3186d7fdc86202a081af2d3a9464). I now copied them from their arXiv pages.
