### 2025-10-07 #8253 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
